# docs/openapi/oauth-user-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arkham Build OAuth and User API",
    "version": "1.0.0",
    "description": "OAuth 2.0 authorization-code API for confidential server-side clients. Client credentials are accepted only in form bodies. Native apps must forward authorization codes to a confidential backend; direct token exchange from an installed app is unsupported."
  },
...
  "servers": [
    {
      "url": "https://api.arkham.build"
    }
  ],
...
  "paths": {
...
    "/v2/user/decks/manifest": {
      "get": {
        "operationId": "getOAuthDeckManifest",
        "tags": ["User decks"],
        "summary": "Get the user's deck manifest",
        "description": "Requires decks:read. Manifests that include ArkhamDB perform an upstream conditional sync and return an opaque snapshot token for consistent batch reads. ArkhamDB failure produces a partial 200 response with account decks and providers.arkhamdb.available=false.",
        "security": [
          {
            "OAuthBearer": ["decks:read"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Optionally include only one provider",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckManifestQuery/properties/source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source-aware deck manifest, possibly partial when ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeckManifestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid source filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/batch": {
      "post": {
        "operationId": "getOAuthDeckBatch",
        "tags": ["User decks"],
        "summary": "Get a batch of decks",
        "description": "Requires decks:read. Accepts at most 250 provider-specific targets and preserves request order. ArkhamDB targets require the snapshot token returned by the manifest endpoint. The whole request fails if any target or the selected snapshot is unavailable.",
        "security": [
          {
            "OAuthBearer": ["decks:read"]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDeckBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Requested decks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeckBatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Batch request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "404": {
            "description": "A deck was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "409": {
            "description": "The selected ArkhamDB snapshot is no longer available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/user/decks/{source}": {
      "post": {
        "operationId": "createOAuthDeck",
        "tags": ["User decks"],
        "summary": "Create a deck",
        "description": "Requires decks:write. Additional request properties, including server-owned deck fields, are ignored. Account creates generate a UUID; ArkhamDB chooses its ID.",
        "security": [
          {
            "OAuthBearer": ["decks:write"]
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "description": "Deck provider: account or arkhamdb",
            "schema": {
              "$ref": "#/components/schemas/OAuthDeckSource"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDeckWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created deck",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthDeck"
                }
              }
            }
          },
          "400": {
            "description": "Deck input is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is unusable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge",
                "schema": {
                  "type": "string",
                  "examples": ["Bearer"]
                }
              }
            }
          },
          "403": {
            "description": "Token lacks decks:write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge, including the required scope when absent",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer error=\"insufficient_scope\", scope=\"decks:write\""
                  ]
                }
              }
            }
          },
          "503": {
            "description": "ArkhamDB is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthUserError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OAuthDeckBatchRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "arkhamdbSyncToken": {
            "description": "Opaque token selecting the exact ArkhamDB snapshot returned by the manifest, or null when ArkhamDB was not included or available",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "decks": {
            "description": "At most 250 source-and-ID targets",
            "maxItems": 250,
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string",
                      "const": "account"
                    },
                    "id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    }
                  },
                  "required": ["source", "id"],
                  "additionalProperties": false
                },
                {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string",
                      "const": "arkhamdb"
                    },
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": ["source", "id"],
                  "additionalProperties": false
                }
              ]
            }
          }
        },
        "required": ["decks"],
        "additionalProperties": false
      },
...
      "OAuthDeckManifestResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "arkhamdbSyncToken": {
            "description": "Opaque token selecting the exact ArkhamDB snapshot returned by the manifest, or null when ArkhamDB was not included or available",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "providers": {
            "type": "object",
            "properties": {
              "account": {
                "type": "object",
                "properties": {
                  "available": {
                    "type": "boolean"
                  }
                },
                "required": ["available"],
                "additionalProperties": false
              },
              "arkhamdb": {
                "type": "object",
                "properties": {
                  "available": {
                    "type": "boolean"
                  }
                },
                "required": ["available"],
                "additionalProperties": false
              }
            },
            "required": ["account", "arkhamdb"],
            "additionalProperties": false
          },
          "decks": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string",
                      "const": "account"
                    },
                    "id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    }
                  },
                  "required": ["source", "id", "updatedAt", "version"],
                  "additionalProperties": false
                },
                {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string",
                      "const": "arkhamdb"
                    },
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    }
                  },
                  "required": ["source", "id", "updatedAt", "version"],
                  "additionalProperties": false
                }
              ]
            }
          }
        },
        "required": ["version", "arkhamdbSyncToken", "providers", "decks"],
        "additionalProperties": false
      },
...
      "OAuthDeckSource": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "description": "Deck provider. Account IDs are strings; ArkhamDB IDs are positive integers in JSON request bodies.",
        "type": "string",
        "enum": ["account", "arkhamdb"]
      },
...
      "OAuthUserError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "invalid_token",
              "insufficient_scope",
              "account_banned",
              "invalid_request",
              "not_found",
              "conflict",
              "upstream_unavailable"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["error", "message"],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "OAuthBearer": {
        "type": "oauth2",
        "description": "Opaque access token sent as Authorization: Bearer ab_at_.... Tokens are issued only to confidential server-side clients.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.arkham.build/v2/oauth/authorize",
            "tokenUrl": "https://api.arkham.build/v2/oauth/token",
            "scopes": {
              "profile:read": "Read the stable account ID and username",
              "decks:read": "Read account and connected ArkhamDB decks",
              "decks:write": "Create, replace, and upgrade decks; implies decks:read",
              "decks:delete": "Delete decks and history; implies decks:write and decks:read"
            }
          }
        }
      }
    }
  }
}